requirements.txt
Pytest==5.3.5
Django==3.1
djangorestframework==3.11
django-debug-toolbar==3.2.2